# Avoid excessive neighbor discovery in NoiseAdaptiveLayout