--- a/qiskit/transpiler/passes/layout/noise_adaptive_layout.py
+++ b/qiskit/transpiler/passes/layout/noise_adaptive_layout.py
@@ -65,6 +65,18 @@ class NoiseAdaptiveLayout(AnalysisPass):
         """
         super().__init__()
         self.backend_prop = backend_prop
+        self.swap_graph = rx.PyDiGraph()
+        self.cx_reliability = {}
+        self.readout_reliability = {}
+        self.available_hw_qubits = []
+        self.gate_list = []
+        self.gate_reliability = {}
+        self.swap_reliabs = {}
+        self.prog_graph = rx.PyGraph()
+        self.prog_neighbors = {}
+        self.qarg_to_id = {}
+        self.pending_program_edges = []
+        self.prog2hw = {}
 
     def _initialize_backend_prop(self):
         """Extract readout and CNOT errors and compute swap costs."""