<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add a release note file using reno if the change needs to be
  documented in the release notes.
⚠️ If your pull request fixes an open issue, please link to the issue.

- [ ] I have added the tests to cover my changes.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the CONTRIBUTING document.
-->

### Summary

This commit fixes a performance regression introduced in #5117 when
there were a large number of gates in the circuit being run through the
pass. The overhead of calling adj() a significant number of times was
adding a runtime penalty. Its not clear what is causing this with
retworkx but my assumption is it is related to going back and forth
between rust and python. This commit mitigates most of this overhead by
locally caching the list of neighbors each time we call
`_select_best_remaining_qubit()`.


### Details and comments