@mtreinish 2020-09-29
This was flagged as a regression in the benchmark runs: https://qiskit.github.io/qiskit/#mapping_passes.PassBenchmarks.time_noise_adaptive_layout?machine=dedicated-benchmarking-softlayer-baremetal&os=Linux%204.15.0-46-generic&ram=16GB&p-n_qubits=20&p-depth=1024&commits=3ed4d182

Running those locally with this applied yields:
```
Benchmarks that have improved:

       before           after         ratio
     [a90e18e9]       [383a176e]
     <noise-adaptive-cache-neighbors~1>       <noise-adaptive-cache-neighbors>
-      11.7±0.6ms      6.38±0.08ms     0.55  mapping_passes.PassBenchmarks.time_noise_adaptive_layout(5, 1024)
-        84.0±1ms         39.4±1ms     0.47  mapping_passes.PassBenchmarks.time_noise_adaptive_layout(20, 1024)
-      57.4±0.5ms       25.6±0.2ms     0.45  mapping_passes.PassBenchmarks.time_noise_adaptive_layout(14, 1024)

Benchmarks that have stayed the same:

       before           after         ratio
     [a90e18e9]       [383a176e]
     <noise-adaptive-cache-neighbors~1>       <noise-adaptive-cache-neighbors>
             130M             130M     1.00  mapping_passes.PassBenchmarks.peakmem_noise_adaptive_layout(5, 1024)
             156M             156M     1.00  mapping_passes.PassBenchmarks.peakmem_noise_adaptive_layout(14, 1024)
             179M             179M     1.00  mapping_passes.PassBenchmarks.peakmem_noise_adaptive_layout(20, 1024)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```